Restructure claims component directory
