+++ apps/admin-portal/src/components/claims/edit/external-dialect/edit-external-claims.tsx
@@ -16,19 +16,19 @@
  * under the License.
  */
 
-import { AlertLevels, ClaimDialect, ExternalClaim } from "../../../models";
-import { ClaimsList, ExternalClaimsSearch, ListType } from "../../../components";
+import { AlertLevels, ClaimDialect, ExternalClaim } from "../../../../models";
+import { ClaimsList, ExternalClaimsSearch, ListType } from "../../..";
 import { Divider, DropdownProps, PaginationProps } from "semantic-ui-react";
-import { filterList, sortList } from "../../../utils";
-import { getADialect, getAllExternalClaims } from "../../../api";
+import { filterList, sortList } from "../../../../utils";
+import { getADialect, getAllExternalClaims } from "../../../../api";
 import React, { useEffect, useState } from "react";
 
 import { addAlert } from "@wso2is/core/store";
-import { AddExternalClaims } from "../add";
-import { DEFAULT_USER_LIST_ITEM_LIMIT } from "../../../constants";
-import { EmptyPlaceholder } from "../../../components/shared";
-import { EmptyPlaceholderIllustrations } from "../../../configs";
-import { ListLayout } from "../../../layouts";
+import { AddExternalClaims } from "../../add";
+import { DEFAULT_USER_LIST_ITEM_LIMIT } from "../../../../constants";
+import { EmptyPlaceholder } from "../../../shared";
+import { EmptyPlaceholderIllustrations } from "../../../../configs";
+import { ListLayout } from "../../../../layouts";
 import { useDispatch } from "react-redux";
 
 interface EditExternalClaimsPropsInterface {
