+++ apps/admin-portal/src/components/claims/edit/local-claim/index.ts
@@ -16,5 +16,6 @@
 * under the License.
 */
 
-export * from "./local-claim";
-export * from "./external-dialect";
+export * from "./edit-additional-properties-local-claims";
+export * from "./edit-basic-details-local-claims";
+export * from "./edit-mapped-attributes-local-claims";
